move appStore to root view
