+++ CounterApp/App/AppCore.swift
@@ -52,18 +52,3 @@ let appReducer = Reducer<AppState, AppAction, AppEnviroment>.combine(
         environment: \.lock
     )
 ).debug()
-
-let appStore = Store(
-    initialState: AppState(
-        counter: CounterState(),
-        lock: LockState(
-            counters: [
-                CounterState(count: 0),
-                CounterState(count: 5),
-                CounterState(count: 9)
-            ]
-        )
-    ),
-    reducer: appReducer,
-    environment: AppEnviroment()
-)
